+++ src/FlatSelect/MultiSelect.md
@@ -1,0 +1,6 @@
+```jsx
+<div>
+	<h1>Максимально упрощённый пример MultiSelect'а</h1>
+	<MultiSelect />
+</div>
+```
