+++ src/FlatSelect/index.ts
@@ -2,3 +2,4 @@ export * from "./FlatSelect";
 export * from "./components";
 export * from "./utils";
 export * from "./types";
+export { MultiSelect as MultiSelectExample } from "./MultiSelect";
